전기 벽 및 게임 시스템 구축
돈, 점수 시스템, 폰트 추가

+++ Assets/02.Scripts/04.Projectile/Bullet.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Bullet : MonoBehaviour
 {
-    private float _damage;
+    private int _damage;
     
     private float _timer;
 
@@ -41,7 +41,7 @@ public class Bullet : MonoBehaviour
     }
 
 
-    public void SetDamage(float damage)
+    public void SetDamage(int damage)
     {
         _damage = damage;
     }
